[VW-316] Move triage into a folder to mirror vex/ layout
agent/triage.ts -> agent/triage/index.ts, agent/triage-context.ts ->
agent/triage/context.ts. Matches the vex/ structure; the public import
(@/features/inbox/agent/triage) is unchanged via index.ts.
